requirements.txt
thop 
onnx
tables  
tqdm 
python-dotenv
pandas 
wandb 
opencv-python
fire 
matplotlib 
lightning
tonic 
sinabs>=2.0
dv-processing 